requirements.txt
Django==2.2.10
djangorestframework==3.11.0
psycopg2==2.8.4
pytz==2019.3
Pillow==6.2.1
sqlparse
django-bootstrap4               # https://pypi.org/project/django-bootstrap4/
django-font-awesome             # https://github.com/benbacardi/django-font-awesome
django-widget-tweaks            # https://github.com/jazzband/django-widget-tweaks
Jinja2                          # https://github.com/noirbizarre/jinja2
viberbot==1.0.11                # https://viber.github.io/docs/api/python-bot-api/
django-summernote==0.8.11.6     # https://github.com/summernote/summernote
django-rest-auth==0.9.5         # https://github.com/Tivix/django-rest-auth
django-cors-headers==3.2.1      # https://github.com/adamchainz/django-cors-headers
celery==4.4.0                   # https://github.com/celery/celery
redis==3.3.11                   # https://redis.io/
django-celery-results==1.1.2    # https://github.com/celery/django-celery-results
django_auto_prefetching==0.1.8  # https://github.com/GeeWee/django-auto-prefetching

# Debuging tools
django_silk-4.0.0               # https://github.com/jazzband/django-silk
